Undo `types.ts` move
