+++ src/typescript/frontend/src/lib/store/event/types.ts
@@ -6,9 +6,9 @@ import {
   type DatabaseModels,
 } from "@sdk/indexer-v2/types";
 import { type SubscribeBarsCallback } from "@static/charting_library/datafeed-api";
-import { type LatestBar } from "./frontend/src/lib/store/event/candlestick-bars";
+import { type LatestBar } from "./candlestick-bars";
 import { type WritableDraft } from "immer";
-import { type ClientState, type ClientActions } from "./frontend/src/lib/store/websocket/store";
+import { type ClientState, type ClientActions } from "../websocket/store";
 
 // Aliased to avoid repeating the type names over and over.
 type Swap = DatabaseModels["swap_events"];
