+++ src/typescript/frontend/tsconfig.json
@@ -70,6 +70,5 @@
     "./src/**/*.tsx",
     ".next/types/**/*.ts",
     "./dist/types/**/*.ts",
-    "../types.ts"
   ]
 }
